pyproject.toml
[tool.poetry]
name = "proxy-auth"
version = "0.1.0"
description = ""
authors = ["Trevor Bernard <trevor.s.bernard@gmail.com>"]
readme = "README.md"
packages = [{include = "proxy_auth"}]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.3"
Flask = "^2.2.3"
pydantic = "^1.10.6"
pyyaml = "^6.0"
flask-sqlalchemy = "^3.0.3"
requests = "^2.28.2"
gunicorn = "^20.1.0"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
proxy_auth_dev = "proxy_auth.app:main"